+++ src/__mocks__/data/mockStore.ts
@@ -1,0 +1,18 @@
+import { StoreType } from 'store/store.types';
+import { account } from './storeData/account';
+import { config } from './storeData/config';
+import { loginInfo } from './storeData/loginInfo';
+import { network } from './storeData/network';
+
+export const mockStore: StoreType = {
+  network: {
+    network
+  },
+  account,
+  loginInfo,
+  config,
+  transactions: {},
+  toasts: { customToasts: [], transactionToasts: [] },
+  cache: {},
+  ui: { isSidePanelOpen: false }
+};