+++ src/managers/ToastManager/helpers/tests/mocks/cache.ts
@@ -1,0 +1,15 @@
+import { testAddress } from '__mocks__/accountConfig';
+
+export const cache = {
+  'transaction-bf1ee4d78ccc57cb81c945decd70678cebc5b8b919f8e3c8411f2b19ebdd6a6x':
+    {
+      asset: null,
+      interactor: testAddress,
+      directionLabel: 'From',
+      action: { name: 'Received xEGLD' },
+      amount: '0 xEGLD',
+      hash: 'bf1ee4d78ccc57cb81c945decd70678cebc5b8b919f8e3c8411f2b19ebdd6a6x',
+      status: 'success',
+      link: 'http://devnet-explorer.multiversx.com/transactions/bf1ee4d78ccc57cb81c945decd70678cebc5b8b919f8e3c8411f2b19ebdd6a6x'
+    }
+};