+++ src/store/actions/transactions/transactionStateByStatus.ts
@@ -3,141 +3,110 @@ import {
   TransactionServerStatusesEnum
 } from 'types/enums.types';
 
-export const pendingBatchTransactionsStates = [
+export const pendingBatchTransactionsStates: string[] = [
   TransactionBatchStatusesEnum.sent
 ];
 
-export const successBatchTransactionsStates = [
+export const successBatchTransactionsStates: string[] = [
   TransactionBatchStatusesEnum.success
 ];
 
-export const failBatchTransactionsStates = [
+export const failBatchTransactionsStates: string[] = [
   TransactionBatchStatusesEnum.fail,
   TransactionBatchStatusesEnum.cancelled,
   TransactionBatchStatusesEnum.timedOut
 ];
 
-export const invalidBatchTransactionsStates = [
+export const invalidBatchTransactionsStates: string[] = [
   TransactionBatchStatusesEnum.invalid
 ];
 
-export const timedOutBatchTransactionsStates = [
+export const timedOutBatchTransactionsStates: string[] = [
   TransactionBatchStatusesEnum.timedOut
 ];
 
-export const pendingServerTransactionsStatuses = [
+export const pendingServerTransactionsStatuses: string[] = [
   TransactionServerStatusesEnum.pending
 ];
 
-export const successServerTransactionsStates = [
+export const successServerTransactionsStates: string[] = [
   TransactionServerStatusesEnum.success
 ];
 
-export const failServerTransactionsStates = [
+export const failServerTransactionsStates: string[] = [
   TransactionServerStatusesEnum.fail,
   TransactionServerStatusesEnum.invalid
 ];
 
-export const notExecutedServerTransactionsStates = [
+export const notExecutedServerTransactionsStates: string[] = [
   TransactionServerStatusesEnum.notExecuted
 ];
 
-export function getIsTransactionPending(
-  status?: TransactionServerStatusesEnum | TransactionBatchStatusesEnum
-) {
+export function getIsTransactionPending(status?: string) {
   return (
     status != null &&
-    (isBatchTransactionPending(status as TransactionBatchStatusesEnum) ||
-      isServerTransactionPending(status as TransactionServerStatusesEnum))
+    (isBatchTransactionPending(status) || isServerTransactionPending(status))
   );
 }
 
-export function getIsTransactionSuccessful(
-  status?: TransactionServerStatusesEnum | TransactionBatchStatusesEnum
-) {
+export function getIsTransactionSuccessful(status?: string) {
   return (
     status != null &&
-    (isBatchTransactionSuccessful(status as TransactionBatchStatusesEnum) ||
-      isServerTransactionSuccessful(status as TransactionServerStatusesEnum))
+    (isBatchTransactionSuccessful(status) ||
+      isServerTransactionSuccessful(status))
   );
 }
 
-export function getIsTransactionFailed(
-  status?: TransactionServerStatusesEnum | TransactionBatchStatusesEnum
-) {
+export function getIsTransactionFailed(status?: string) {
   return (
     status != null &&
-    (isBatchTransactionFailed(status as TransactionBatchStatusesEnum) ||
-      isServerTransactionFailed(status as TransactionServerStatusesEnum))
+    (isBatchTransactionFailed(status) || isServerTransactionFailed(status))
   );
 }
 
-export function getIsTransactionNotExecuted(
-  status?: TransactionServerStatusesEnum | TransactionBatchStatusesEnum
-) {
+export function getIsTransactionNotExecuted(status?: string) {
   return (
     status != null &&
-    (isBatchTransactionInvalid(status as TransactionBatchStatusesEnum) ||
-      isServerTransactionNotExecuted(status as TransactionServerStatusesEnum))
+    (isBatchTransactionInvalid(status) ||
+      isServerTransactionNotExecuted(status))
   );
 }
 
-export function getIsTransactionTimedOut(
-  status?: TransactionServerStatusesEnum | TransactionBatchStatusesEnum
-) {
-  return (
-    status != null &&
-    isBatchTransactionTimedOut(status as TransactionBatchStatusesEnum)
-  );
+export function getIsTransactionTimedOut(status?: string) {
+  return status != null && isBatchTransactionTimedOut(status);
 }
 
-export function isBatchTransactionPending(
-  status?: TransactionBatchStatusesEnum
-) {
+export function isBatchTransactionPending(status?: string) {
   return status != null && pendingBatchTransactionsStates.includes(status);
 }
 
-export function isBatchTransactionSuccessful(
-  status: TransactionBatchStatusesEnum
-) {
+export function isBatchTransactionSuccessful(status?: string) {
   return status != null && successBatchTransactionsStates.includes(status);
 }
 
-export function isBatchTransactionFailed(status: TransactionBatchStatusesEnum) {
+export function isBatchTransactionFailed(status?: string) {
   return status != null && failBatchTransactionsStates.includes(status);
 }
 
-export function isBatchTransactionInvalid(
-  status: TransactionBatchStatusesEnum
-) {
+export function isBatchTransactionInvalid(status?: string) {
   return status != null && invalidBatchTransactionsStates.includes(status);
 }
 
-export function isBatchTransactionTimedOut(
-  status?: TransactionBatchStatusesEnum
-) {
+export function isBatchTransactionTimedOut(status?: string) {
   return status != null && timedOutBatchTransactionsStates.includes(status);
 }
 
-export function isServerTransactionPending(
-  status?: TransactionServerStatusesEnum
-) {
+export function isServerTransactionPending(status?: string) {
   return status != null && pendingServerTransactionsStatuses.includes(status);
 }
-export function isServerTransactionSuccessful(
-  status: TransactionServerStatusesEnum
-) {
+export function isServerTransactionSuccessful(status?: string) {
   return status != null && successServerTransactionsStates.includes(status);
 }
 
-export function isServerTransactionFailed(
-  status: TransactionServerStatusesEnum
-) {
+export function isServerTransactionFailed(status?: string) {
   return status != null && failServerTransactionsStates.includes(status);
 }
 
-export function isServerTransactionNotExecuted(
-  status: TransactionServerStatusesEnum
-) {
+export function isServerTransactionNotExecuted(status?: string) {
   return status != null && notExecutedServerTransactionsStates.includes(status);
 }